Rename and add handleclick
